Create a separate crate for the Prisma Schema AST definition
Scope: only the definition of the AST contents. This does not include
the parser or the reformatter yet, because we need to disentangle error
handling first.

Goal: the datamodel connector interface is in its own crate and cannot
depend on datamodel/core (because that would be a circular dependency).
We want to use AST types in the datamodel connectors, so we have to
split them off, like it was done for the dml crate when native types
were implemented.

Additionally,

- Splitting this into its own crate will make dependencies between
  different components of the schema parser clearer.
- The AST API being public means it can be more strictly enforced and
  visibly reflected in cargo docs.

This PR is mostly just moves existing code, with a few minor
improvements and new docs.
