+++ libs/datamodel/core/Cargo.toml
@@ -9,6 +9,7 @@ datamodel-connector = { path = "../connectors/datamodel-connector" }
 dml = { path = "../connectors/dml" }
 mongodb-datamodel-connector = { path = "../connectors/mongodb-datamodel-connector" }
 prisma-value = { path = "../../prisma-value" }
+schema-ast = { path = "../schema-ast" }
 sql-datamodel-connector = { path = "../connectors/sql-datamodel-connector" }
 
 bigdecimal = "0.2"
