+++ libs/datamodel/schema-ast/src/ast/generator_config.rs
@@ -1,4 +1,4 @@
-use super::*;
+use super::{Argument, Comment, Identifier, Span, WithDocumentation, WithIdentifier, WithSpan};
 
 /// A Generator block declaration.
 #[derive(Debug, Clone, PartialEq)]
